requirements.txt
flask
RPi.GPIO
uwsgi
astral